requirements.txt

azure-functions
msal
requests
